+++ paker/exceptions.py
@@ -1,5 +1,3 @@
-
-
 class PakerImportError(Exception):
 
     def __init__(self, message):